% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampBasemap.R
\name{ampBasemap}
\alias{ampBasemap}
\title{AMP Basemap creation for use in plotting}
\usage{
ampBasemap(
  data,
  coastline,
  aoi_buffer = 1000,
  facility = NA,
  analysis = FALSE,
  cageExtents = FALSE,
  leases = FALSE,
  sampleLocations = FALSE
)
}
\arguments{
\item{data}{list of AMP data from ampDataLoad function}

\item{coastline}{sf object from ampMapSetup function}

\item{aoi_buffer}{distance in meters for how far from sample locations to
buffer for map creation}
}
\value{
ggplot2 plot of sf objects
}
\description{
AMP Basemap creation for use in plotting
}
\examples{
ampBasemap()
}
