% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampInterpolation.R
\name{ampInterpolation}
\alias{ampInterpolation}
\title{AMP Bathymetric interpolation}
\usage{
ampInterpolation(data, x, y, z, cellsize, crs = 32620)
}
\arguments{
\item{data}{sf dataframe of data with coordinates variables included}

\item{x}{longitude}

\item{y}{latitude}

\item{z}{parameter column name to be interpolated}

\item{cellsize}{grid cell size in meters}
}
\value{
dataframe of kriged valuess
}
\description{
AMP Bathymetric interpolation
}
\examples{
ampInterpolation()
}
