% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampKrige.R
\name{ampKrige}
\alias{ampKrige}
\title{AMP parameter Universal Kriging}
\usage{
ampKrige(dataframe, z, cellsize, crs = 32620)
}
\arguments{
\item{dataframe}{sf dataframe of data with coordinates variables included}

\item{z}{parameter column name to be interpolated}

\item{cellsize}{grid cell size in meters}

\item{crs}{ending CRS, should keep as 32620 unless you have reasons not to}
}
\value{
dataframe of kriged valuess
}
\description{
AMP parameter Universal Kriging
}
\examples{
ampKrige()
}
