% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampKrigePlot.R
\name{ampKrigePlot}
\alias{ampKrigePlot}
\title{Title}
\usage{
ampKrigePlot(
  krigedData,
  contour = TRUE,
  contour_interval = 10,
  crs = 32620,
  colourTheme = "plasma"
)
}
\arguments{
\item{contour}{}
}
\description{
Title
}
