% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampMapSetup.R
\name{ampMapSetup}
\alias{ampMapSetup}
\title{AMP map setup}
\usage{
ampMapSetup(
  path = "data/input/coastline/",
  country = "canada",
  province = "N.B.",
  dl_coastline = TRUE
)
}
\arguments{
\item{path}{character string; location to save downloads to}

\item{country}{`canada` or `usa`}

\item{dl_coastline}{logical; if TRUE will download the Statistics Canada coastline file}

\item{dl_us_coastline}{logical; if TRUE will download the NOAA coastline file}
}
\value{
Downloads Canadian or USA coastline shapefile data to path
}
\description{
AMP map setup
}
\examples{
ampMapSetup()

}
