% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampDataLoad.R
\name{\%>\%}
\alias{\%>\%}
\alias{ampDataLoad}
\title{Data Wrangling for Aquaculture Monitoring Program}
\usage{
ampDataLoad(
  file,
  excelSheets = c("si", "m", "s", "gs", "dpab"),
  facilityNum = FALSE
)
}
\arguments{
\item{excelSheets}{the sheet name from the excel spreadsheet of data}

\item{facilityNum}{the facilityNumberInProject from the spreadsheet for the
facility you are interested in}

\item{x}{AMP data file for a specific region and year}
}
\value{
a list for each aquaculture facility, as well as a `combined` list of
  all of the data
}
\description{
Data Wrangling for Aquaculture Monitoring Program
}
\examples{
# some examples if you want to highlight the usage in the package
}
