% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampBasemap.R, R/ampDataLoad.R, R/ampKrige.R
\name{\%>\%}
\alias{\%>\%}
\alias{ampBasemap}
\alias{ampDataLoad}
\alias{ampKrige}
\title{AMP Basemap creation for use in plotting}
\usage{
ampBasemap(
  data,
  coastline,
  aoi_buffer = 1000,
  insetMap = FALSE,
  facility = NA,
  cageExtents = FALSE
)

ampDataLoad(file, excelSheets = c("si", "m", "s", "gs", "dpab"))

ampKrige(dataframe, z, cellsize, crs = 32620)
}
\arguments{
\item{data}{list of AMP data from ampDataLoad function}

\item{coastline}{sf object from ampMapSetup function}

\item{aoi_buffer}{distance in meters for how far from sample locations to
buffer for map creation}

\item{excelSheets}{the sheet name from the excel spreadsheet of data}

\item{dataframe}{sf dataframe of data with coordinates variables included}

\item{z}{parameter column name to be interpolated}

\item{cellsize}{grid cell size in meters}

\item{facilityNum}{the facilityNumberInProject from the spreadsheet for the
facility you are interested in}

\item{x}{longitude}

\item{y}{latitude}
}
\value{
ggplot2 plot of sf objects

a list for each aquaculture facility, as well as a `combined` list of
  all of the data

dataframe of kriged valuess
}
\description{
AMP Basemap creation for use in plotting

Data Wrangling for Aquaculture Monitoring Program

AMP parameter Universal Kriging
}
\examples{
ampBasemap()
# some examples if you want to highlight the usage in the package
ampKrige()
}
